Segregate rockets to separate file
